Final Fix

+++ inc/Blockchain/PKCertChain/Certificate.h
@@ -5,7 +5,7 @@
 #include <ctime>
 
 struct certificate {
-    uint16_t nodeID;
+    uint8_t nodeID;
     std::vector<uint8_t> publicKey;
     std::vector<uint8_t> signature;
     time_t notBefore;
